% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_cuda_methods.R
\name{print.ridge_cuda}
\alias{print.ridge_cuda}
\title{Print Method for ridge_cuda Objects}
\usage{
\method{print}{ridge_cuda}(x, ...)
}
\arguments{
\item{x}{A `ridge_cuda` object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object `x`.
}
\description{
Print Method for ridge_cuda Objects
}
\keyword{internal}
