% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_cuda_methods.R
\name{summary.ridge_cuda}
\alias{summary.ridge_cuda}
\title{Summary Method for ridge_cuda Objects}
\usage{
\method{summary}{ridge_cuda}(object, ...)
}
\arguments{
\item{object}{A `ridge_cuda` object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A list object of class `summary.ridge_cuda`.
}
\description{
Summary Method for ridge_cuda Objects
}
\keyword{internal}
